Updates files organization
