+++ UpcomingMovies/Helpers/Utils/ProgressHud/Configuration/HudConfigurationProtocol.swift
@@ -1,0 +1,20 @@
+//
+//  HudConfigurationProtocol.swift
+//  UpcomingMovies
+//
+//  Created by Alonso on 16/11/21.
+//  Copyright © 2021 Alonso. All rights reserved.
+//
+
+import UIKit
+
+protocol HudConfigurationProtocol {
+
+    var backgroundColor: UIColor { get }
+    var hudColor: UIColor { get }
+    var activityIndicatorColor: UIColor { get }
+
+    var textColor: UIColor { get }
+    var textFont: UIFont { get }
+
+}
